修复 socket.close 关闭错误。修复 libev recv 错误。

+++ fibjs/src/MemoryStream.cpp
@@ -155,7 +155,7 @@ result_t MemoryStream::seek(double offset, int32_t whence)
 	if (offset < 0)
 		offset = 0;
 	else if (offset > sz)
-		offset = sz;
+		offset = (double)sz;
 
 	m_buffer.seekg((int64_t) offset);
 	m_buffer.seekp((int64_t) offset);
